refactor(gov): rearrange code to maintain easily
