+++ contracts/pylon/gov/src/querier/poll.rs
@@ -1,0 +1,110 @@
+use cosmwasm_std::{CanonicalAddr, StdResult, Storage};
+use cosmwasm_storage::ReadonlyBucket;
+use pylon_token::common::OrderBy;
+use pylon_token::gov::{PollStatus, VoterInfo};
+
+use crate::state::poll::{
+    poll_indexer_r, poll_r, poll_voter_r, tmp_poll_id_r, tmp_poll_id_w, Poll,
+};
+
+const MAX_LIMIT: u32 = 30;
+const DEFAULT_LIMIT: u32 = 10;
+
+pub fn store_tmp_poll_id(storage: &mut dyn Storage, poll_id: u64) -> StdResult<()> {
+    tmp_poll_id_w(storage).save(&poll_id)
+}
+
+pub fn read_tmp_poll_id(storage: &dyn Storage) -> StdResult<u64> {
+    tmp_poll_id_r(storage).load()
+}
+
+pub fn polls<'a>(
+    storage: &'a dyn Storage,
+    filter: Option<PollStatus>,
+    start_after: Option<u64>,
+    limit: Option<u32>,
+    order_by: Option<OrderBy>,
+) -> StdResult<Vec<Poll>> {
+    let limit = limit.unwrap_or(DEFAULT_LIMIT).min(MAX_LIMIT) as usize;
+    let (start, end, order_by) = match order_by {
+        Some(OrderBy::Asc) => (calc_range_start(start_after), None, OrderBy::Asc),
+        _ => (None, calc_range_end(start_after), OrderBy::Desc),
+    };
+
+    if let Some(status) = filter {
+        let poll_indexer: ReadonlyBucket<'a, bool> = poll_indexer_r(storage, &status);
+
+        poll_indexer
+            .range(start.as_deref(), end.as_deref(), order_by.into())
+            .take(limit)
+            .map(|item| {
+                let (k, _) = item?;
+                poll_r(storage).load(&k)
+            })
+            .collect()
+    } else {
+        let polls: ReadonlyBucket<'a, Poll> = poll_r(storage);
+
+        polls
+            .range(start.as_deref(), end.as_deref(), order_by.into())
+            .take(limit)
+            .map(|item| {
+                let (_, v) = item?;
+                Ok(v)
+            })
+            .collect()
+    }
+}
+
+pub fn poll_voters<'a>(
+    storage: &'a dyn Storage,
+    poll_id: u64,
+    start_after: Option<CanonicalAddr>,
+    limit: Option<u32>,
+    order_by: Option<OrderBy>,
+) -> StdResult<Vec<(CanonicalAddr, VoterInfo)>> {
+    let limit = limit.unwrap_or(DEFAULT_LIMIT).min(MAX_LIMIT) as usize;
+    let (start, end, order_by) = match order_by {
+        Some(OrderBy::Asc) => (calc_range_start_addr(start_after), None, OrderBy::Asc),
+        _ => (None, calc_range_end_addr(start_after), OrderBy::Desc),
+    };
+
+    let poll_voters: ReadonlyBucket<'a, VoterInfo> = poll_voter_r(storage, poll_id);
+
+    poll_voters
+        .range(start.as_deref(), end.as_deref(), order_by.into())
+        .take(limit)
+        .map(|item| {
+            let (k, v) = item?;
+            Ok((CanonicalAddr::from(k), v))
+        })
+        .collect()
+}
+
+// this will set the first key after the provided key, by appending a 1 byte
+fn calc_range_start(start_after: Option<u64>) -> Option<Vec<u8>> {
+    start_after.map(|id| {
+        let mut v = id.to_be_bytes().to_vec();
+        v.push(1);
+        v
+    })
+}
+
+// this will set the first key after the provided key, by appending a 1 byte
+fn calc_range_end(start_after: Option<u64>) -> Option<Vec<u8>> {
+    start_after.map(|id| id.to_be_bytes().to_vec())
+}
+
+// this will set the first key after the provided key, by appending a 1 byte
+fn calc_range_start_addr(start_after: Option<CanonicalAddr>) -> Option<Vec<u8>> {
+    start_after.map(|address| {
+        let mut v = address.as_slice().to_vec();
+        v.push(1);
+        v
+    })
+}
+
+// this will set the first key after the provided key, by appending a 1 byte
+fn calc_range_end_addr(start_after: Option<CanonicalAddr>) -> Option<Vec<u8>> {
+    start_after.map(|address| address.as_slice().to_vec())
+}
