+++ contracts/pylon/gov/src/state/mod.rs
@@ -1,0 +1,5 @@
+pub mod bank;
+pub mod config;
+pub mod poll;
+#[allow(clippy::module_inception)]
+pub mod state;
