+++ contracts/pylon/gov/src/testing/constants.rs
@@ -1,0 +1,11 @@
+pub const VOTING_TOKEN: &str = "voting_token";
+pub const TEST_CREATOR: &str = "creator";
+pub const TEST_VOTER: &str = "voter1";
+pub const TEST_VOTER_2: &str = "voter2";
+pub const TEST_VOTER_3: &str = "voter3";
+pub const DEFAULT_QUORUM: u64 = 30u64;
+pub const DEFAULT_THRESHOLD: u64 = 50u64;
+pub const DEFAULT_VOTING_PERIOD: u64 = 20000u64;
+pub const DEFAULT_FIX_PERIOD: u64 = 10u64;
+pub const DEFAULT_TIMELOCK_PERIOD: u64 = 10000u64;
+pub const DEFAULT_PROPOSAL_DEPOSIT: u128 = 10000000000u128;
