+++ contracts/pylon/gov/src/testing/test_poll_end.rs
@@ -1,0 +1,1120 @@
+use cosmwasm_std::testing::{mock_env, mock_info, MOCK_CONTRACT_ADDR};
+use cosmwasm_std::{
+    attr, coins, from_binary, to_binary, Addr, Api, CosmosMsg, SubMsg, Uint128, WasmMsg,
+};
+use cw20::{Cw20ExecuteMsg, Cw20ReceiveMsg};
+use pylon_token::common::OrderBy;
+use pylon_token::gov::{
+    Cw20HookMsg, ExecuteMsg, PollExecuteMsg, PollMsg, PollResponse, PollStatus, PollsResponse,
+    QueryMsg, StakerResponse, VoteOption, VoterInfo, VotersResponse,
+};
+use terraswap::querier::query_token_balance;
+
+use crate::contract;
+use crate::error::ContractError;
+use crate::state::bank::bank_r;
+use crate::state::poll::poll_voter_r;
+use crate::testing::assert::{
+    assert_cast_vote_success, assert_create_poll_result, assert_stake_tokens_result,
+};
+use crate::testing::constants::*;
+use crate::testing::message::create_poll_msg;
+use crate::testing::mock_querier::mock_dependencies;
+use crate::testing::utils::{mock_env_height, mock_instantiate};
+
+#[test]
+fn end_poll() {
+    const POLL_START_HEIGHT: u64 = 1000;
+    const POLL_ID: u64 = 1;
+    let stake_amount = 1000;
+
+    let mut deps = mock_dependencies(&coins(1000, VOTING_TOKEN));
+    mock_instantiate(deps.as_mut());
+
+    let mut creator_env = mock_env_height(POLL_START_HEIGHT, 10000);
+    let mut creator_info = mock_info(VOTING_TOKEN, &coins(2, VOTING_TOKEN));
+
+    let exec_msg_bz = to_binary(&Cw20ExecuteMsg::Burn {
+        amount: Uint128::new(123),
+    })
+    .unwrap();
+
+    let exec_msg_bz2 = to_binary(&Cw20ExecuteMsg::Burn {
+        amount: Uint128::new(12),
+    })
+    .unwrap();
+
+    let exec_msg_bz3 = to_binary(&Cw20ExecuteMsg::Burn {
+        amount: Uint128::new(1),
+    })
+    .unwrap();
+
+    //add three messages with different order
+    let execute_msgs: Vec<PollExecuteMsg> = vec![
+        PollExecuteMsg {
+            order: 3u64,
+            contract: VOTING_TOKEN.to_string(),
+            msg: exec_msg_bz3.clone(),
+        },
+        PollExecuteMsg {
+            order: 2u64,
+            contract: VOTING_TOKEN.to_string(),
+            msg: exec_msg_bz2.clone(),
+        },
+        PollExecuteMsg {
+            order: 1u64,
+            contract: VOTING_TOKEN.to_string(),
+            msg: exec_msg_bz.clone(),
+        },
+    ];
+
+    let msg = create_poll_msg(
+        "test".to_string(),
+        "test".to_string(),
+        None,
+        Some(execute_msgs),
+    );
+
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap();
+
+    assert_create_poll_result(
+        1,
+        creator_env.block.height + DEFAULT_VOTING_PERIOD,
+        TEST_CREATOR,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from((stake_amount + DEFAULT_PROPOSAL_DEPOSIT) as u128),
+        )],
+    )]);
+
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER.to_string(),
+        amount: Uint128::from(stake_amount as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    let execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+    assert_stake_tokens_result(
+        stake_amount,
+        DEFAULT_PROPOSAL_DEPOSIT,
+        stake_amount,
+        1,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    let msg = ExecuteMsg::Poll(PollMsg::CastVote {
+        poll_id: 1,
+        vote: VoteOption::Yes,
+        amount: Uint128::from(stake_amount),
+    });
+    let env = mock_env_height(POLL_START_HEIGHT, 10000);
+    let info = mock_info(TEST_VOTER, &[]);
+    let execute_res = contract::execute(deps.as_mut(), env, info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "cast_vote"),
+            attr("poll_id", POLL_ID.to_string().as_str()),
+            attr("amount", "1000"),
+            attr("voter", TEST_VOTER),
+            attr("vote_option", "yes"),
+        ]
+    );
+
+    // not in passed status
+    let msg = ExecuteMsg::Poll(PollMsg::Execute { poll_id: 1 });
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap_err();
+    match execute_res {
+        ContractError::PollNotPassed {} => (),
+        _ => panic!("DO NOT ENTER HERE"),
+    }
+
+    creator_info.sender = Addr::unchecked(TEST_CREATOR);
+    creator_env.block.height += DEFAULT_VOTING_PERIOD;
+
+    let msg = ExecuteMsg::Poll(PollMsg::End { poll_id: 1 });
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "end_poll"),
+            attr("poll_id", "1"),
+            attr("rejected_reason", ""),
+            attr("passed", "true"),
+        ]
+    );
+    assert_eq!(
+        execute_res.messages,
+        vec![SubMsg::new(CosmosMsg::Wasm(WasmMsg::Execute {
+            contract_addr: VOTING_TOKEN.to_string(),
+            msg: to_binary(&Cw20ExecuteMsg::Transfer {
+                recipient: TEST_CREATOR.to_string(),
+                amount: Uint128::from(DEFAULT_PROPOSAL_DEPOSIT),
+            })
+            .unwrap(),
+            funds: vec![],
+        }))]
+    );
+
+    // End poll will withdraw deposit balance
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from(stake_amount as u128),
+        )],
+    )]);
+
+    // timelock_period has not expired
+    let msg = ExecuteMsg::Poll(PollMsg::Execute { poll_id: 1 });
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap_err();
+    match execute_res {
+        ContractError::TimelockNotExpired {} => (),
+        _ => panic!("DO NOT ENTER HERE"),
+    }
+
+    creator_env.block.height += DEFAULT_TIMELOCK_PERIOD;
+    let msg = ExecuteMsg::Poll(PollMsg::Execute { poll_id: 1 });
+    let execute_res =
+        contract::execute(deps.as_mut(), creator_env.clone(), creator_info, msg).unwrap();
+    assert_eq!(
+        execute_res.messages,
+        vec![SubMsg::reply_on_error(
+            CosmosMsg::Wasm(WasmMsg::Execute {
+                contract_addr: creator_env.contract.address.to_string(),
+                msg: to_binary(&ExecuteMsg::Poll(PollMsg::ExecuteMsgs { poll_id: 1 })).unwrap(),
+                funds: vec![],
+            }),
+            1
+        )]
+    );
+
+    let msg = ExecuteMsg::Poll(PollMsg::ExecuteMsgs { poll_id: 1 });
+    let contract_info = mock_info(MOCK_CONTRACT_ADDR, &[]);
+    let execute_res = contract::execute(deps.as_mut(), creator_env, contract_info, msg).unwrap();
+    assert_eq!(
+        execute_res.messages,
+        vec![
+            SubMsg::new(CosmosMsg::Wasm(WasmMsg::Execute {
+                contract_addr: VOTING_TOKEN.to_string(),
+                msg: exec_msg_bz,
+                funds: vec![],
+            })),
+            SubMsg::new(CosmosMsg::Wasm(WasmMsg::Execute {
+                contract_addr: VOTING_TOKEN.to_string(),
+                msg: exec_msg_bz2,
+                funds: vec![],
+            })),
+            SubMsg::new(CosmosMsg::Wasm(WasmMsg::Execute {
+                contract_addr: VOTING_TOKEN.to_string(),
+                msg: exec_msg_bz3,
+                funds: vec![],
+            }))
+        ]
+    );
+    assert_eq!(
+        execute_res.attributes,
+        vec![attr("action", "execute_poll"), attr("poll_id", "1"),]
+    );
+
+    // Query executed polls
+    let res = contract::query(
+        deps.as_ref(),
+        mock_env(),
+        QueryMsg::Polls {
+            filter: Some(PollStatus::Passed),
+            start_after: None,
+            limit: None,
+            order_by: None,
+        },
+    )
+    .unwrap();
+    let response: PollsResponse = from_binary(&res).unwrap();
+    assert_eq!(response.polls.len(), 0);
+
+    let res = contract::query(
+        deps.as_ref(),
+        mock_env(),
+        QueryMsg::Polls {
+            filter: Some(PollStatus::InProgress),
+            start_after: None,
+            limit: None,
+            order_by: None,
+        },
+    )
+    .unwrap();
+    let response: PollsResponse = from_binary(&res).unwrap();
+    assert_eq!(response.polls.len(), 0);
+
+    let res = contract::query(
+        deps.as_ref(),
+        mock_env(),
+        QueryMsg::Polls {
+            filter: Some(PollStatus::Executed),
+            start_after: None,
+            limit: None,
+            order_by: Some(OrderBy::Desc),
+        },
+    )
+    .unwrap();
+    let response: PollsResponse = from_binary(&res).unwrap();
+    assert_eq!(response.polls.len(), 1);
+
+    // voter info must be deleted
+    let res = contract::query(
+        deps.as_ref(),
+        mock_env(),
+        QueryMsg::Voters {
+            poll_id: 1u64,
+            start_after: None,
+            limit: None,
+            order_by: None,
+        },
+    )
+    .unwrap();
+    let response: VotersResponse = from_binary(&res).unwrap();
+    assert_eq!(response.voters.len(), 0);
+
+    // staker locked token must be disappeared
+    let res = contract::query(
+        deps.as_ref(),
+        mock_env(),
+        QueryMsg::Staker {
+            address: TEST_VOTER.to_string(),
+        },
+    )
+    .unwrap();
+    let response: StakerResponse = from_binary(&res).unwrap();
+    assert_eq!(
+        response,
+        StakerResponse {
+            balance: Uint128::from(stake_amount),
+            share: Uint128::from(stake_amount),
+            locked_balance: vec![]
+        }
+    );
+
+    // But the data is still in the store
+    let voter_addr_raw = deps.api.addr_canonicalize(TEST_VOTER).unwrap();
+    let voter = poll_voter_r(&deps.storage, 1u64)
+        .load(voter_addr_raw.as_slice())
+        .unwrap();
+    assert_eq!(
+        voter,
+        VoterInfo {
+            vote: VoteOption::Yes,
+            balance: Uint128::from(stake_amount),
+        }
+    );
+
+    let token_manager = bank_r(&deps.storage)
+        .load(voter_addr_raw.as_slice())
+        .unwrap();
+    assert_eq!(
+        token_manager.locked_balance,
+        vec![(
+            1u64,
+            VoterInfo {
+                vote: VoteOption::Yes,
+                balance: Uint128::from(stake_amount),
+            }
+        )]
+    );
+}
+
+#[test]
+fn end_poll_with_controlled_quorum() {
+    const POLL_START_HEIGHT: u64 = 1000;
+    const POLL_ID: u64 = 1;
+    let stake_amount = 1000;
+
+    let mut deps = mock_dependencies(&coins(1000, VOTING_TOKEN));
+    mock_instantiate(deps.as_mut());
+
+    let mut creator_env = mock_env_height(POLL_START_HEIGHT, 10000);
+    let mut creator_info = mock_info(VOTING_TOKEN, &coins(2, VOTING_TOKEN));
+
+    let exec_msg_bz = to_binary(&Cw20ExecuteMsg::Burn {
+        amount: Uint128::new(123),
+    })
+    .unwrap();
+
+    //add two messages
+    let execute_msgs: Vec<PollExecuteMsg> = vec![
+        PollExecuteMsg {
+            order: 1u64,
+            contract: VOTING_TOKEN.to_string(),
+            msg: exec_msg_bz.clone(),
+        },
+        PollExecuteMsg {
+            order: 2u64,
+            contract: VOTING_TOKEN.to_string(),
+            msg: exec_msg_bz,
+        },
+    ];
+
+    let msg = create_poll_msg(
+        "test".to_string(),
+        "test".to_string(),
+        None,
+        Some(execute_msgs),
+    );
+
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap();
+
+    assert_create_poll_result(
+        1,
+        creator_env.block.height + DEFAULT_VOTING_PERIOD,
+        TEST_CREATOR,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from((stake_amount + DEFAULT_PROPOSAL_DEPOSIT) as u128),
+        )],
+    )]);
+
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER.to_string(),
+        amount: Uint128::from(stake_amount as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    let execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+    assert_stake_tokens_result(
+        stake_amount,
+        DEFAULT_PROPOSAL_DEPOSIT,
+        stake_amount,
+        1,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    let msg = ExecuteMsg::Poll(PollMsg::CastVote {
+        poll_id: 1,
+        vote: VoteOption::Yes,
+        amount: Uint128::from(stake_amount),
+    });
+    let env = mock_env_height(POLL_START_HEIGHT, 10000);
+    let info = mock_info(TEST_VOTER, &[]);
+    let execute_res = contract::execute(deps.as_mut(), env, info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "cast_vote"),
+            attr("poll_id", POLL_ID.to_string().as_str()),
+            attr("amount", "1000"),
+            attr("voter", TEST_VOTER),
+            attr("vote_option", "yes"),
+        ]
+    );
+
+    creator_env.block.height += DEFAULT_VOTING_PERIOD - 10;
+
+    // send SnapshotPoll
+    let fix_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        ExecuteMsg::Poll(PollMsg::Snapshot { poll_id: 1 }),
+    )
+    .unwrap();
+
+    assert_eq!(
+        fix_res.attributes,
+        vec![
+            attr("action", "snapshot_poll"),
+            attr("poll_id", "1"),
+            attr("staked_amount", stake_amount.to_string().as_str()),
+        ]
+    );
+
+    // staked amount get increased 10 times
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from(((10 * stake_amount) + DEFAULT_PROPOSAL_DEPOSIT) as u128),
+        )],
+    )]);
+
+    //cast another vote
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER_2.to_string(),
+        amount: Uint128::from(8 * stake_amount as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    let _execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+
+    let msg = ExecuteMsg::Poll(PollMsg::CastVote {
+        poll_id: 1,
+        vote: VoteOption::Yes,
+        amount: Uint128::from(8 * stake_amount),
+    });
+    let env = mock_env_height(creator_env.block.height, 10000);
+    let info = mock_info(TEST_VOTER_2, &[]);
+    let execute_res = contract::execute(deps.as_mut(), env, info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "cast_vote"),
+            attr("poll_id", POLL_ID.to_string().as_str()),
+            attr("amount", "8000"),
+            attr("voter", TEST_VOTER_2),
+            attr("vote_option", "yes"),
+        ]
+    );
+
+    creator_info.sender = Addr::unchecked(TEST_CREATOR);
+    creator_env.block.height += 10;
+
+    // quorum must reach
+    let msg = ExecuteMsg::Poll(PollMsg::End { poll_id: 1 });
+    let execute_res = contract::execute(deps.as_mut(), creator_env, creator_info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "end_poll"),
+            attr("poll_id", "1"),
+            attr("rejected_reason", ""),
+            attr("passed", "true"),
+        ]
+    );
+    assert_eq!(
+        execute_res.messages,
+        vec![SubMsg::new(CosmosMsg::Wasm(WasmMsg::Execute {
+            contract_addr: VOTING_TOKEN.to_string(),
+            msg: to_binary(&Cw20ExecuteMsg::Transfer {
+                recipient: TEST_CREATOR.to_string(),
+                amount: Uint128::from(DEFAULT_PROPOSAL_DEPOSIT),
+            })
+            .unwrap(),
+            funds: vec![],
+        }))]
+    );
+
+    let res = contract::query(deps.as_ref(), mock_env(), QueryMsg::Poll { poll_id: 1 }).unwrap();
+    let value: PollResponse = from_binary(&res).unwrap();
+    assert_eq!(
+        stake_amount,
+        value.total_balance_at_end_poll.unwrap().u128()
+    );
+
+    assert_eq!(value.yes_votes.u128(), 9 * stake_amount);
+
+    // actual staked amount is 10 times bigger than staked amount
+    let actual_staked_weight = query_token_balance(
+        &deps.as_ref().querier,
+        Addr::unchecked(VOTING_TOKEN),
+        Addr::unchecked(MOCK_CONTRACT_ADDR),
+    )
+    .unwrap()
+    .checked_sub(Uint128::from(DEFAULT_PROPOSAL_DEPOSIT))
+    .unwrap();
+
+    assert_eq!(actual_staked_weight.u128(), (10 * stake_amount))
+}
+
+#[test]
+fn end_poll_zero_quorum() {
+    let mut deps = mock_dependencies(&coins(1000, VOTING_TOKEN));
+    mock_instantiate(deps.as_mut());
+
+    let mut creator_env = mock_env_height(1000, 10000);
+    let mut creator_info = mock_info(VOTING_TOKEN, &[]);
+
+    let execute_msgs: Vec<PollExecuteMsg> = vec![PollExecuteMsg {
+        order: 1u64,
+        contract: VOTING_TOKEN.to_string(),
+        msg: to_binary(&Cw20ExecuteMsg::Burn {
+            amount: Uint128::new(123),
+        })
+        .unwrap(),
+    }];
+
+    let msg = create_poll_msg(
+        "test".to_string(),
+        "test".to_string(),
+        None,
+        Some(execute_msgs),
+    );
+
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap();
+    assert_create_poll_result(
+        1,
+        creator_env.block.height + DEFAULT_VOTING_PERIOD,
+        TEST_CREATOR,
+        execute_res,
+        deps.as_ref(),
+    );
+    let stake_amount = 100;
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from(100u128 + DEFAULT_PROPOSAL_DEPOSIT),
+        )],
+    )]);
+
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER.to_string(),
+        amount: Uint128::from(stake_amount as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+
+    let msg = ExecuteMsg::Poll(PollMsg::End { poll_id: 1 });
+    creator_info.sender = Addr::unchecked(TEST_CREATOR);
+    creator_env.block.height += DEFAULT_VOTING_PERIOD;
+
+    let execute_res = contract::execute(deps.as_mut(), creator_env, creator_info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "end_poll"),
+            attr("poll_id", "1"),
+            attr("rejected_reason", "Quorum not reached"),
+            attr("passed", "false"),
+        ]
+    );
+
+    assert_eq!(execute_res.messages.len(), 0usize);
+
+    // Query rejected polls
+    let res = contract::query(
+        deps.as_ref(),
+        mock_env(),
+        QueryMsg::Polls {
+            filter: Some(PollStatus::Rejected),
+            start_after: None,
+            limit: None,
+            order_by: Some(OrderBy::Desc),
+        },
+    )
+    .unwrap();
+    let response: PollsResponse = from_binary(&res).unwrap();
+    assert_eq!(response.polls.len(), 1);
+
+    let res = contract::query(
+        deps.as_ref(),
+        mock_env(),
+        QueryMsg::Polls {
+            filter: Some(PollStatus::InProgress),
+            start_after: None,
+            limit: None,
+            order_by: None,
+        },
+    )
+    .unwrap();
+    let response: PollsResponse = from_binary(&res).unwrap();
+    assert_eq!(response.polls.len(), 0);
+
+    let res = contract::query(
+        deps.as_ref(),
+        mock_env(),
+        QueryMsg::Polls {
+            filter: Some(PollStatus::Passed),
+            start_after: None,
+            limit: None,
+            order_by: None,
+        },
+    )
+    .unwrap();
+    let response: PollsResponse = from_binary(&res).unwrap();
+    assert_eq!(response.polls.len(), 0);
+}
+
+#[test]
+fn end_poll_quorum_rejected() {
+    let mut deps = mock_dependencies(&coins(100, VOTING_TOKEN));
+    mock_instantiate(deps.as_mut());
+
+    let msg = create_poll_msg("test".to_string(), "test".to_string(), None, None);
+    let mut creator_env = mock_env();
+    let mut creator_info = mock_info(VOTING_TOKEN, &[]);
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap();
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "create_poll"),
+            attr("creator", TEST_CREATOR),
+            attr("poll_id", "1"),
+            attr("end_height", "32345"),
+        ]
+    );
+
+    let stake_amount = 100;
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from(100u128 + DEFAULT_PROPOSAL_DEPOSIT),
+        )],
+    )]);
+
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER.to_string(),
+        amount: Uint128::from(stake_amount as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    let execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+    assert_stake_tokens_result(
+        stake_amount,
+        DEFAULT_PROPOSAL_DEPOSIT,
+        stake_amount,
+        1,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    let msg = ExecuteMsg::Poll(PollMsg::CastVote {
+        poll_id: 1,
+        vote: VoteOption::Yes,
+        amount: Uint128::from(10u128),
+    });
+    let info = mock_info(TEST_VOTER, &[]);
+    let execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "cast_vote"),
+            attr("poll_id", "1"),
+            attr("amount", "10"),
+            attr("voter", TEST_VOTER),
+            attr("vote_option", "yes"),
+        ]
+    );
+
+    let msg = ExecuteMsg::Poll(PollMsg::End { poll_id: 1 });
+
+    creator_info.sender = Addr::unchecked(TEST_CREATOR);
+    creator_env.block.height += DEFAULT_VOTING_PERIOD;
+
+    let execute_res = contract::execute(deps.as_mut(), creator_env, creator_info, msg).unwrap();
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "end_poll"),
+            attr("poll_id", "1"),
+            attr("rejected_reason", "Quorum not reached"),
+            attr("passed", "false"),
+        ]
+    );
+}
+
+#[test]
+fn end_poll_quorum_rejected_nothing_staked() {
+    let mut deps = mock_dependencies(&coins(100, VOTING_TOKEN));
+    mock_instantiate(deps.as_mut());
+
+    let msg = create_poll_msg("test".to_string(), "test".to_string(), None, None);
+    let mut creator_env = mock_env();
+    let mut creator_info = mock_info(VOTING_TOKEN, &[]);
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap();
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "create_poll"),
+            attr("creator", TEST_CREATOR),
+            attr("poll_id", "1"),
+            attr("end_height", "32345"),
+        ]
+    );
+
+    let msg = ExecuteMsg::Poll(PollMsg::End { poll_id: 1 });
+
+    creator_info.sender = Addr::unchecked(TEST_CREATOR);
+    creator_env.block.height += DEFAULT_VOTING_PERIOD;
+
+    let execute_res = contract::execute(deps.as_mut(), creator_env, creator_info, msg).unwrap();
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "end_poll"),
+            attr("poll_id", "1"),
+            attr("rejected_reason", "Quorum not reached"),
+            attr("passed", "false"),
+        ]
+    );
+}
+
+#[test]
+fn end_poll_nay_rejected() {
+    let voter1_stake = 100;
+    let voter2_stake = 1000;
+    let mut deps = mock_dependencies(&[]);
+    mock_instantiate(deps.as_mut());
+
+    let mut creator_env = mock_env();
+    let mut creator_info = mock_info(VOTING_TOKEN, &coins(2, VOTING_TOKEN));
+
+    let msg = create_poll_msg("test".to_string(), "test".to_string(), None, None);
+
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap();
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "create_poll"),
+            attr("creator", TEST_CREATOR),
+            attr("poll_id", "1"),
+            attr("end_height", "32345"),
+        ]
+    );
+
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from((voter1_stake + DEFAULT_PROPOSAL_DEPOSIT) as u128),
+        )],
+    )]);
+
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER.to_string(),
+        amount: Uint128::from(voter1_stake as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    let execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+    assert_stake_tokens_result(
+        voter1_stake,
+        DEFAULT_PROPOSAL_DEPOSIT,
+        voter1_stake,
+        1,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from((voter1_stake + voter2_stake + DEFAULT_PROPOSAL_DEPOSIT) as u128),
+        )],
+    )]);
+
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER_2.to_string(),
+        amount: Uint128::from(voter2_stake as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    let execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+    assert_stake_tokens_result(
+        voter1_stake + voter2_stake,
+        DEFAULT_PROPOSAL_DEPOSIT,
+        voter2_stake,
+        1,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    let info = mock_info(TEST_VOTER_2, &[]);
+    let msg = ExecuteMsg::Poll(PollMsg::CastVote {
+        poll_id: 1,
+        vote: VoteOption::No,
+        amount: Uint128::from(voter2_stake),
+    });
+    let execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+    assert_cast_vote_success(TEST_VOTER_2, voter2_stake, 1, VoteOption::No, execute_res);
+
+    let msg = ExecuteMsg::Poll(PollMsg::End { poll_id: 1 });
+
+    creator_info.sender = Addr::unchecked(TEST_CREATOR);
+    creator_env.block.height += DEFAULT_VOTING_PERIOD;
+    let execute_res = contract::execute(deps.as_mut(), creator_env, creator_info, msg).unwrap();
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "end_poll"),
+            attr("poll_id", "1"),
+            attr("rejected_reason", "Threshold not reached"),
+            attr("passed", "false"),
+        ]
+    );
+}
+
+#[test]
+fn fails_end_poll_before_end_height() {
+    let mut deps = mock_dependencies(&[]);
+    mock_instantiate(deps.as_mut());
+
+    let env = mock_env_height(0, 10000);
+    let info = mock_info(VOTING_TOKEN, &[]);
+
+    let msg = create_poll_msg("test".to_string(), "test".to_string(), None, None);
+
+    let execute_res = contract::execute(deps.as_mut(), env, info, msg).unwrap();
+    assert_create_poll_result(
+        1,
+        DEFAULT_VOTING_PERIOD,
+        TEST_CREATOR,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    let res = contract::query(deps.as_ref(), mock_env(), QueryMsg::Poll { poll_id: 1 }).unwrap();
+    let value: PollResponse = from_binary(&res).unwrap();
+    assert_eq!(DEFAULT_VOTING_PERIOD, value.end_height);
+
+    let msg = ExecuteMsg::Poll(PollMsg::End { poll_id: 1 });
+    let env = mock_env_height(0, 10000);
+    let info = mock_info(TEST_CREATOR, &[]);
+    let execute_res = contract::execute(deps.as_mut(), env, info, msg);
+
+    match execute_res {
+        Ok(_) => panic!("Must return error"),
+        Err(ContractError::PollVotingPeriod {}) => (),
+        Err(e) => panic!("Unexpected error: {:?}", e),
+    }
+}
+
+#[test]
+fn fails_end_poll_quorum_inflation_without_snapshot_poll() {
+    const POLL_START_HEIGHT: u64 = 1000;
+    const POLL_ID: u64 = 1;
+    let stake_amount = 1000;
+
+    let mut deps = mock_dependencies(&coins(1000, VOTING_TOKEN));
+    mock_instantiate(deps.as_mut());
+
+    let mut creator_env = mock_env_height(POLL_START_HEIGHT, 10000);
+    let mut creator_info = mock_info(VOTING_TOKEN, &coins(2, VOTING_TOKEN));
+
+    let exec_msg_bz = to_binary(&Cw20ExecuteMsg::Burn {
+        amount: Uint128::new(123),
+    })
+    .unwrap();
+
+    //add two messages
+    let execute_msgs: Vec<PollExecuteMsg> = vec![
+        PollExecuteMsg {
+            order: 1u64,
+            contract: VOTING_TOKEN.to_string(),
+            msg: exec_msg_bz.clone(),
+        },
+        PollExecuteMsg {
+            order: 2u64,
+            contract: VOTING_TOKEN.to_string(),
+            msg: exec_msg_bz,
+        },
+    ];
+
+    let msg = create_poll_msg(
+        "test".to_string(),
+        "test".to_string(),
+        None,
+        Some(execute_msgs),
+    );
+
+    let execute_res = contract::execute(
+        deps.as_mut(),
+        creator_env.clone(),
+        creator_info.clone(),
+        msg,
+    )
+    .unwrap();
+
+    assert_create_poll_result(
+        1,
+        creator_env.block.height + DEFAULT_VOTING_PERIOD,
+        TEST_CREATOR,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from((stake_amount + DEFAULT_PROPOSAL_DEPOSIT) as u128),
+        )],
+    )]);
+
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER.to_string(),
+        amount: Uint128::from(stake_amount as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    let execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+    assert_stake_tokens_result(
+        stake_amount,
+        DEFAULT_PROPOSAL_DEPOSIT,
+        stake_amount,
+        1,
+        execute_res,
+        deps.as_ref(),
+    );
+
+    let msg = ExecuteMsg::Poll(PollMsg::CastVote {
+        poll_id: 1,
+        vote: VoteOption::Yes,
+        amount: Uint128::from(stake_amount),
+    });
+    let env = mock_env_height(POLL_START_HEIGHT, 10000);
+    let info = mock_info(TEST_VOTER, &[]);
+    let execute_res = contract::execute(deps.as_mut(), env, info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "cast_vote"),
+            attr("poll_id", POLL_ID.to_string().as_str()),
+            attr("amount", "1000"),
+            attr("voter", TEST_VOTER),
+            attr("vote_option", "yes"),
+        ]
+    );
+
+    creator_env.block.height += DEFAULT_VOTING_PERIOD - 10;
+
+    // did not SnapshotPoll
+
+    // staked amount get increased 10 times
+    deps.querier.with_token_balances(&[(
+        &VOTING_TOKEN.to_string(),
+        &[(
+            &MOCK_CONTRACT_ADDR.to_string(),
+            &Uint128::from(((10 * stake_amount) + DEFAULT_PROPOSAL_DEPOSIT) as u128),
+        )],
+    )]);
+
+    //cast another vote
+    let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
+        sender: TEST_VOTER_2.to_string(),
+        amount: Uint128::from(8 * stake_amount as u128),
+        msg: to_binary(&Cw20HookMsg::StakeVotingTokens {}).unwrap(),
+    });
+
+    let info = mock_info(VOTING_TOKEN, &[]);
+    let _execute_res = contract::execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+
+    // another voter cast a vote
+    let msg = ExecuteMsg::Poll(PollMsg::CastVote {
+        poll_id: 1,
+        vote: VoteOption::Yes,
+        amount: Uint128::from(stake_amount),
+    });
+    let env = mock_env_height(creator_env.block.height, 10000);
+    let info = mock_info(TEST_VOTER_2, &[]);
+    let execute_res = contract::execute(deps.as_mut(), env, info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "cast_vote"),
+            attr("poll_id", POLL_ID.to_string().as_str()),
+            attr("amount", "1000"),
+            attr("voter", TEST_VOTER_2),
+            attr("vote_option", "yes"),
+        ]
+    );
+
+    creator_info.sender = Addr::unchecked(TEST_CREATOR);
+    creator_env.block.height += 10;
+
+    // quorum must reach
+    let msg = ExecuteMsg::Poll(PollMsg::End { poll_id: 1 });
+    let execute_res = contract::execute(deps.as_mut(), creator_env, creator_info, msg).unwrap();
+
+    assert_eq!(
+        execute_res.attributes,
+        vec![
+            attr("action", "end_poll"),
+            attr("poll_id", "1"),
+            attr("rejected_reason", "Quorum not reached"),
+            attr("passed", "false"),
+        ]
+    );
+
+    let res = contract::query(deps.as_ref(), mock_env(), QueryMsg::Poll { poll_id: 1 }).unwrap();
+    let value: PollResponse = from_binary(&res).unwrap();
+    assert_eq!(
+        10 * stake_amount,
+        value.total_balance_at_end_poll.unwrap().u128()
+    );
+}
